Move typval code to typval.c.
Better separate code operating of typvals from code evaluating vimL.
