+++ src/nvim/eval.h
@@ -2,6 +2,7 @@
 #define NVIM_EVAL_H
 
 #include "nvim/profile.h"
+#include "nvim/typval.h"
 
 /* Defines for Vim variables.  These must match vimvars[] in eval.c! */
 enum {
